requirements.txt
fastapi
uvicorn[standard]
requests
PyJWT
python-dotenv